Add CatalogBlockItemModuleFactory
